Put database related methods into own file.
